// gen/ts/open_splunk/v1/lookup_api.ts
export enum LookupSortBy {
  LOOKUP_SORT_BY_UNSPECIFIED = 0,
  LOOKUP_SORT_BY_NAME = 1,
  LOOKUP_SORT_BY_CREATED_AT = 2,
  LOOKUP_SORT_BY_UPDATED_AT = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface GetLookupResponse {
  lookup: Lookup | undefined;
}

/** POST /api/v1/knowledge/lookups/list */
export interface ListLookupsRequest {
  page: PageRequest | undefined;
  appId?: string | undefined;
  stateFilters: LookupState[];
  textFilter?: string | undefined;
  sortBy: LookupSortBy;
  sortDirection: SortDirection;
}

export interface ListLookupsResponse {
  lookups: Lookup[];
  page: PageResponse | undefined;
}
// ...
export interface DeleteLookupResponse {
  lookupId: string;
  version: bigint;
}
// ...
export const GetLookupResponse: MessageFns<GetLookupResponse> = {
  encode(message: GetLookupResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.lookup !== undefined) {
      Lookup.encode(message.lookup, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetLookupResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetLookupResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.lookup = Lookup.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetLookupResponse {
    return { lookup: isSet(object.lookup) ? Lookup.fromJSON(object.lookup) : undefined };
  },

  toJSON(message: GetLookupResponse): unknown {
    const obj: any = {};
    if (message.lookup !== undefined) {
      obj.lookup = Lookup.toJSON(message.lookup);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetLookupResponse>, I>>(base?: I): GetLookupResponse {
    return GetLookupResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetLookupResponse>, I>>(object: I): GetLookupResponse {
    const message = createBaseGetLookupResponse();
    message.lookup = (object.lookup !== undefined && object.lookup !== null)
      ? Lookup.fromPartial(object.lookup)
      : undefined;
    return message;
  },
};
// ...
export const ListLookupsResponse: MessageFns<ListLookupsResponse> = {
  encode(message: ListLookupsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.lookups) {
      Lookup.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.page !== undefined) {
      PageResponse.encode(message.page, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ListLookupsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseListLookupsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.lookups.push(Lookup.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.page = PageResponse.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ListLookupsResponse {
    return {
      lookups: globalThis.Array.isArray(object?.lookups) ? object.lookups.map((e: any) => Lookup.fromJSON(e)) : [],
      page: isSet(object.page) ? PageResponse.fromJSON(object.page) : undefined,
    };
  },

  toJSON(message: ListLookupsResponse): unknown {
    const obj: any = {};
    if (message.lookups?.length) {
      obj.lookups = message.lookups.map((e) => Lookup.toJSON(e));
    }
    if (message.page !== undefined) {
      obj.page = PageResponse.toJSON(message.page);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ListLookupsResponse>, I>>(base?: I): ListLookupsResponse {
    return ListLookupsResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ListLookupsResponse>, I>>(object: I): ListLookupsResponse {
    const message = createBaseListLookupsResponse();
    message.lookups = object.lookups?.map((e) => Lookup.fromPartial(e)) || [];
    message.page = (object.page !== undefined && object.page !== null)
      ? PageResponse.fromPartial(object.page)
      : undefined;
    return message;
  },
};
// ...
export const DeleteLookupResponse: MessageFns<DeleteLookupResponse> = {
  encode(message: DeleteLookupResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.lookupId !== "") {
      writer.uint32(10).string(message.lookupId);
    }
    if (message.version !== 0n) {
      if (BigInt.asUintN(64, message.version) !== message.version) {
        throw new globalThis.Error("value provided for field message.version of type uint64 too large");
      }
      writer.uint32(16).uint64(message.version);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteLookupResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteLookupResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.lookupId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.version = reader.uint64() as bigint;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DeleteLookupResponse {
    return {
      lookupId: isSet(object.lookupId)
        ? globalThis.String(object.lookupId)
        : isSet(object.lookup_id)
        ? globalThis.String(object.lookup_id)
        : "",
      version: isSet(object.version) ? BigInt(object.version) : 0n,
    };
  },

  toJSON(message: DeleteLookupResponse): unknown {
    const obj: any = {};
    if (message.lookupId !== "") {
      obj.lookupId = message.lookupId;
    }
    if (message.version !== 0n) {
      obj.version = message.version.toString();
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DeleteLookupResponse>, I>>(base?: I): DeleteLookupResponse {
    return DeleteLookupResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DeleteLookupResponse>, I>>(object: I): DeleteLookupResponse {
    const message = createBaseDeleteLookupResponse();
    message.lookupId = object.lookupId ?? "";
    message.version = (object.version !== undefined && object.version !== null) ? BigInt(object.version) : 0n;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string; value: unknown } ? { $case: T["$case"]; value?: DeepPartial<T["value"]> }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}